requirements.txt
torch>=2.2.0
transformers>=4.44.0
datasets>=2.20.0
numpy>=1.26.0
scipy>=1.11.0
scikit-learn>=1.3.0
networkx>=3.2
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.66.0
pyyaml>=6.0.0
jupyter>=1.0.0
ipykernel>=6.0.0
streamlit>=1.28.0
plotly>=5.17.0
